% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{execute_rest_query}
\alias{execute_rest_query}
\title{Execute a DAX query against a specified PowerBI Dataset using the REST API
endpoint.}
\usage{
execute_rest_query(workspace, dataset, query, powerbi_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset to execute query against}

\item{query}{DAX query to execute}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing results of query
}
\description{
Execute a DAX query against a specified PowerBI Dataset using the REST API
endpoint.
}
