% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_table}
\alias{get_dataset_table}
\title{Download a specified dataset table into R using either the XMLA or REST API
endpoints. This is a convenience wrapper around the \code{execute_xmla_query()}
and \code{execute_rest_query()} functions.}
\usage{
get_dataset_table(
  workspace,
  dataset,
  table,
  method = "XMLA",
  powerbi_token = get_powerbi_token(application_id =
    "23d8f6bd-1eb0-4cc2-a08c-7bf525c67bcd")
)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset containing table}

\item{table}{Name of the table to download}

\item{method}{The API to use for downloading the table. Valid values are "XMLA" (the default) and "REST"}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing downloaded table
}
\description{
Download a specified dataset table into R using either the XMLA or REST API
endpoints. This is a convenience wrapper around the \code{execute_xmla_query()}
and \code{execute_rest_query()} functions.
}
