% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_table_xmla}
\alias{get_dataset_table_xmla}
\title{Download a specified dataset table into R using the XMLA API endpoint. This
is a convenience wrapper around the the \code{execute_xmla_query()} function.}
\usage{
get_dataset_table_xmla(
  workspace,
  dataset,
  table,
  powerbi_token = get_powerbi_token(application_id =
    "23d8f6bd-1eb0-4cc2-a08c-7bf525c67bcd")
)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataset}{Name of the dataset containing table}

\item{table}{Name of the table to download}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing downloaded table
}
\description{
Download a specified dataset table into R using the XMLA API endpoint. This
is a convenience wrapper around the the \code{execute_xmla_query()} function.
}
