% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_powerbi_token.R
\name{get_powerbi_token}
\alias{get_powerbi_token}
\title{Generate an Azure authentication token for use with PowerBI}
\usage{
get_powerbi_token(
  tenant_id = getOption("powerbi.tenant_id", "common"),
  application_id = getOption("powerbi.application_id",
    "1950a258-227b-4e31-a9cf-717495945fc2"),
  resource_id = getOption("powerbi.tenant_id",
    "https://analysis.windows.net/powerbi/api"),
  ...
)
}
\arguments{
\item{tenant_id}{Your organisation's tenant identifier, defaults to 'common'}

\item{application_id}{The application identifier with access to PowerBI,
defaults to '1950a258-227b-4e31-a9cf-717495945fc2' (Azure PowerShell Modules)}

\item{resource_id}{The URL/GUID of the target resource, defaults to 'https://analysis.windows.net/powerbi/api'}

\item{...}{Additional arguments to be passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token()}}, #nolint
such as 'use_cache' or 'auth_type'}
}
\value{
An Azure token to authenticate connections.
}
\description{
Generate an Azure authentication token for use with PowerBI
}
