% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{get_table_sas_key}
\alias{get_table_sas_key}
\title{Request the Azure Blob SaS key needed to download a specified dataflow table}
\usage{
get_table_sas_key(workspace, dataflow, table, powerbi_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataflow}{Name of the dataflow containing tables}

\item{table}{Name of the table to download}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing table metadata and SaS key
}
\description{
Request the Azure Blob SaS key needed to download a specified dataflow table
}
