% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{list_dataflow_tables}
\alias{list_dataflow_tables}
\title{Request metadata for all tables in a specified dataflow in specified workspace}
\usage{
list_dataflow_tables(workspace, dataflow, powerbi_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflow}

\item{dataflow}{Name of the dataflow containing tables}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing the names, column names, and column types, for all tables in dataflow
}
\description{
Request metadata for all tables in a specified dataflow in specified workspace
}
