% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataflows.R
\name{list_dataflows}
\alias{list_dataflows}
\title{Request metadata for all dataflows in specified workspace}
\usage{
list_dataflows(workspace, powerbi_token)
}
\arguments{
\item{workspace}{Name of the workspace containing dataflows}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing the names, GUIDs, and descriptions for all dataflows in workspace
}
\description{
Request metadata for all dataflows in specified workspace
}
