% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Request metadata for all datasets in specified workspace}
\usage{
list_datasets(workspace, powerbi_token)
}
\arguments{
\item{workspace}{Name of the workspace containing datasets}

\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing the names, GUIDs, and descriptions for all datasets in workspace
}
\description{
Request metadata for all datasets in specified workspace
}
