% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces.R
\name{list_workspaces}
\alias{list_workspaces}
\title{Request metadata for all Workspaces in Tenant}
\usage{
list_workspaces(powerbi_token)
}
\arguments{
\item{powerbi_token}{AzureAuth token object generated by \code{get_powerbi_token()}}
}
\value{
DataFrame containing the names, GUIDs, and capacity GUIDs for all workspaces in tenant.
}
\description{
Request metadata for all Workspaces in Tenant
}
