% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerbiapi-package.R
\docType{package}
\name{PowerBIAPI-package}
\alias{PowerBIAPI-package}
\alias{PowerBIAPI}
\title{The PowerBIAPI package.}
\description{
An R interface to REST APIs for PowerBI Dataflows and Datasets
}
